Consider arrays as values
When customizing a payload, anyway array should be overridden because
one might want to have an empty array, or one different from the one in
the payload.  Having to manually removes items after the generation is
not going to look pretty.

Bonus: Made the tests more bdd-ish and made sure the original payload
was not modifiable